Refactor request handling
This is a thought experiment to see what it would look like to avoid all of those `if` statements in the request handling.  It requires some knowledge of some things most people don't use in Java (enum methods, really?!) and adds a lot of additional lines of code.  The method signatures for the abstract methods are also a little funky to handle the needs of each action.  That said, we could readily move that request handler to another file to lower the line count here and pull that functionality into its own logical unit.  Enum methods are pretty useful in cases like this and if it's part of the language, then it's fair game as long as it's not *too* clever (right?).  And the method signatures could be condensed if we wrapped them in a type or did a more full-featured refactoring.

The benefits of this approach is encapsulation, (subjective) beauty, and (potential) testability.  All of the logic to match and execute in each state is encapsulated in the enum definition.  We reduce the maintenance burden by making restricting how much context a developer needs to know about each potential state and simplifying code paths (no branching).  Finally, the handler itself is an object that could be pulled out and unit tested.  My favorite thing is that adding new states does not require any knowledge of the surrounding code.  We can just add a new enum and implement two methods and not worry about any other state.