Add custom dimension fields for analytics (#59)
Add custom dimension fields for analytics

Add cd1-cd5 to pass custom data to google analytics. Some data is
duplicated but that is to allow easier filtering. The only piece
of data not readily available was teamName, to easily access this
data it has been added to the bot info object.